Changed Entry.cs to Globals.cs, made it global, and created a new Entry.cs with the starting scene
