+++ project.godot
@@ -15,6 +15,10 @@ run/main_scene="uid://bhn3okwreagx3"
 config/features=PackedStringArray("4.4", "C#", "Forward Plus")
 config/icon="res://icon.svg"
 
+[autoload]
+
+Globals="*res://scripts/Globals.cs"
+
 [display]
 
 window/size/viewport_width=1280
